sentinel zuul refactor
