+++ spring-cloud-alibaba-sentinel-zuul/src/main/java/org/springframework/cloud/alibaba/sentinel/zuul/SentinelZuulAutoConfiguration.java
@@ -16,6 +16,17 @@
 
 package org.springframework.cloud.alibaba.sentinel.zuul;
 
+import static org.springframework.cloud.alibaba.sentinel.zuul.SentinelZuulAutoConfiguration.PREFIX;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.support.DefaultListableBeanFactory;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.cloud.alibaba.sentinel.zuul.listener.FallBackProviderHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+
 import com.alibaba.csp.sentinel.adapter.zuul.fallback.DefaultRequestOriginParser;
 import com.alibaba.csp.sentinel.adapter.zuul.fallback.DefaultUrlCleaner;
 import com.alibaba.csp.sentinel.adapter.zuul.fallback.RequestOriginParser;
@@ -25,17 +36,8 @@ import com.alibaba.csp.sentinel.adapter.zuul.filters.SentinelPostFilter;
 import com.alibaba.csp.sentinel.adapter.zuul.filters.SentinelPreFilter;
 import com.alibaba.csp.sentinel.adapter.zuul.properties.SentinelZuulProperties;
 import com.alibaba.csp.sentinel.util.StringUtil;
-import com.netflix.zuul.ZuulFilter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.support.DefaultListableBeanFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cloud.alibaba.sentinel.zuul.listener.FallBackProviderListener;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
 
-import static org.springframework.cloud.alibaba.sentinel.zuul.SentinelZuulAutoConfiguration.PREFIX;
+import com.netflix.zuul.ZuulFilter;
 
 /**
  * Sentinel Spring Cloud Zuul AutoConfiguration
@@ -46,65 +48,67 @@ import static org.springframework.cloud.alibaba.sentinel.zuul.SentinelZuulAutoCo
 @ConditionalOnProperty(prefix = PREFIX, name = "enabled", havingValue = "true")
 public class SentinelZuulAutoConfiguration {
 
-    @Autowired
-    private Environment environment;
+	@Autowired
+	private Environment environment;
 
-    public static final String PREFIX = "spring.cloud.alibaba.sentinel.zuul";
+	public static final String PREFIX = "spring.cloud.alibaba.sentinel.zuul";
 
-    @Bean
-    public SentinelZuulProperties sentinelZuulProperties() {
-        SentinelZuulProperties properties = new SentinelZuulProperties();
-        String enabledStr = environment.getProperty(PREFIX + "." + "enabled");
-        String preOrderStr = environment.getProperty(PREFIX + "." + "order.pre");
-        String postOrderStr = environment.getProperty(PREFIX + "." + "order.post");
-        String errorOrderStr = environment.getProperty(PREFIX + "." + "order.error");
-        if (StringUtil.isNotEmpty(enabledStr)) {
-            Boolean enabled = Boolean.valueOf(enabledStr);
-            properties.setEnabled(enabled);
-        }
-        if (StringUtil.isNotEmpty(preOrderStr)) {
-            properties.getOrder().setPre(Integer.parseInt(preOrderStr));
-        }
-        if (StringUtil.isNotEmpty(postOrderStr)) {
-            properties.getOrder().setPost(Integer.parseInt(postOrderStr));
-        }
-        if (StringUtil.isNotEmpty(errorOrderStr)) {
-            properties.getOrder().setError(Integer.parseInt(errorOrderStr));
-        }
-        return properties;
-    }
+	@Bean
+	public SentinelZuulProperties sentinelZuulProperties() {
+		SentinelZuulProperties properties = new SentinelZuulProperties();
+		String enabledStr = environment.getProperty(PREFIX + "." + "enabled");
+		String preOrderStr = environment.getProperty(PREFIX + "." + "order.pre");
+		String postOrderStr = environment.getProperty(PREFIX + "." + "order.post");
+		String errorOrderStr = environment.getProperty(PREFIX + "." + "order.error");
+		if (StringUtil.isNotEmpty(enabledStr)) {
+			Boolean enabled = Boolean.valueOf(enabledStr);
+			properties.setEnabled(enabled);
+		}
+		if (StringUtil.isNotEmpty(preOrderStr)) {
+			properties.getOrder().setPre(Integer.parseInt(preOrderStr));
+		}
+		if (StringUtil.isNotEmpty(postOrderStr)) {
+			properties.getOrder().setPost(Integer.parseInt(postOrderStr));
+		}
+		if (StringUtil.isNotEmpty(errorOrderStr)) {
+			properties.getOrder().setError(Integer.parseInt(errorOrderStr));
+		}
+		return properties;
+	}
 
-    @Bean
-    @ConditionalOnMissingBean(UrlCleaner.class)
-    public UrlCleaner urlCleaner(){
-        return new DefaultUrlCleaner();
-    }
+	@Bean
+	@ConditionalOnMissingBean(UrlCleaner.class)
+	public UrlCleaner urlCleaner() {
+		return new DefaultUrlCleaner();
+	}
 
-    @Bean
-    @ConditionalOnMissingBean(RequestOriginParser.class)
-    public RequestOriginParser requestOriginParser(){
-        return new DefaultRequestOriginParser();
-    }
+	@Bean
+	@ConditionalOnMissingBean(RequestOriginParser.class)
+	public RequestOriginParser requestOriginParser() {
+		return new DefaultRequestOriginParser();
+	}
 
-    @Bean
-    public ZuulFilter preFilter(SentinelZuulProperties sentinelZuulProperties,UrlCleaner urlCleaner,
-                                RequestOriginParser requestOriginParser) {
-        return new SentinelPreFilter(sentinelZuulProperties,urlCleaner,requestOriginParser);
-    }
+	@Bean
+	public ZuulFilter preFilter(SentinelZuulProperties sentinelZuulProperties,
+			UrlCleaner urlCleaner, RequestOriginParser requestOriginParser) {
+		return new SentinelPreFilter(sentinelZuulProperties, urlCleaner,
+				requestOriginParser);
+	}
 
-    @Bean
-    public ZuulFilter postFilter(SentinelZuulProperties sentinelZuulProperties) {
-        return new SentinelPostFilter(sentinelZuulProperties);
-    }
+	@Bean
+	public ZuulFilter postFilter(SentinelZuulProperties sentinelZuulProperties) {
+		return new SentinelPostFilter(sentinelZuulProperties);
+	}
 
-    @Bean
-    public ZuulFilter errorFilter(SentinelZuulProperties sentinelZuulProperties) {
-        return new SentinelErrorFilter(sentinelZuulProperties);
-    }
+	@Bean
+	public ZuulFilter errorFilter(SentinelZuulProperties sentinelZuulProperties) {
+		return new SentinelErrorFilter(sentinelZuulProperties);
+	}
 
-    @Bean
-    public FallBackProviderListener fallBackProviderListener(DefaultListableBeanFactory beanFactory) {
-        return new FallBackProviderListener(beanFactory);
-    }
+	@Bean
+	public FallBackProviderHandler fallBackProviderListener(
+			DefaultListableBeanFactory beanFactory) {
+		return new FallBackProviderHandler(beanFactory);
+	}
 
 }
